+++ static/components/f-blocks.js
@@ -1,5 +1,5 @@
 import { asset, push } from "./utils.js";
-import { Field } from "./field.js";
+import { Component } from "./component.js";
 import { init } from "../libs/gutenberg.js";
 
 const styleLink = document.createElement("link");
@@ -9,10 +9,10 @@ document.head.appendChild(styleLink);
 
 customElements.define(
   "f-blocks",
-  class extends Field {
+  class extends Component {
     init() {
       this.classList.add("field");
-      const { schema, value, namePrefix } = this;
+      const { schema, value, namePrefix, isNew } = this;
       const name = `${namePrefix}.${schema.name}`;
       const id = `field_${name}`;
 
@@ -25,7 +25,7 @@ customElements.define(
       const textarea = push(this, "textarea", {
         id,
         name,
-        value,
+        value: isNew ? schema.value : value,
         hidden: true,
       });