+++ static/components/ui.js
@@ -1,3 +1,4 @@
+import "./u-fields.js";
 import "./u-copy.js";
 import "./u-draggable.js";
 import "./u-filter.js";